+++ examples/src/ref_finance.rs
@@ -42,14 +42,16 @@ async fn create_ref(owner: &Account, worker: &Worker<Sandbox>) -> anyhow::Result
             "referral_fee": 1,
         }))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     owner
         .call(ref_finance.id(), "storage_deposit")
         .args_json(serde_json::json!({}))
         .deposit(parse_near!("30 mN"))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     Ok(ref_finance)
 }
@@ -71,20 +73,23 @@ async fn create_wnear(owner: &Account, worker: &Worker<Sandbox>) -> anyhow::Resu
             "total_supply": parse_near!("1,000,000,000 N"),
         }))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     owner
         .call(wnear.id(), "storage_deposit")
         .args_json(serde_json::json!({}))
         .deposit(parse_near!("0.008 N"))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     owner
         .call(wnear.id(), "near_deposit")
         .deposit(parse_near!("200 N"))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     Ok(wnear)
 }
@@ -106,7 +111,8 @@ async fn create_pool_with_liquidity(
         .call("extend_whitelisted_tokens")
         .args_json(serde_json::json!({ "tokens": token_ids }))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     let pool_id: u64 = ref_finance
         .call("add_simple_pool")
@@ -126,7 +132,8 @@ async fn create_pool_with_liquidity(
         }))
         .deposit(1)
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     deposit_tokens(owner, &ref_finance, tokens).await?;
 
@@ -138,7 +145,8 @@ async fn create_pool_with_liquidity(
         }))
         .deposit(parse_near!("1 N"))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     Ok(pool_id)
 }
@@ -158,7 +166,8 @@ async fn deposit_tokens(
             }))
             .deposit(parse_near!("1 N"))
             .transact()
-            .await?;
+            .await?
+            .into_result()?;
 
         owner
             .call(contract_id, "ft_transfer_call")
@@ -170,7 +179,8 @@ async fn deposit_tokens(
             .gas(parse_gas!("200 Tgas") as u64)
             .deposit(1)
             .transact()
-            .await?;
+            .await?
+            .into_result()?;
     }
 
     Ok(())
@@ -193,7 +203,9 @@ async fn create_custom_ft(
             "total_supply": parse_near!("1,000,000,000 N").to_string(),
         }))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
+    ();
 
     Ok(ft)
 }