+++ workspaces/tests/batch_tx.rs
@@ -24,7 +24,8 @@ async fn test_batch_tx() -> anyhow::Result<()> {
             "message": "world_hello",
         })))
         .transact()
-        .await?;
+        .await?
+        .into_result()?;
 
     let status_msg: String = contract
         .call("get_status")